feat: tests for server error (429, 502)
